Wed, Jul 27, 2022, 11:41 AM +00:00

+++ readability/readability.c
@@ -24,7 +24,7 @@ int main(void)
     printf("Sen is %i\n",sentences);
     printf("words is %i\n",words);
     printf("letters is %i\n",letters);
-    double L= (letters / words ) *100.0, S=(sentences/words) *100.0;
+    float L= (letters / words ) *100.0, S=(sentences/words) *100.0;
     printf("L is %f\n",L);
     printf("S is %f\n",S);
     grade=round(0.0588 * L - 0.296 * S - 15.8);
